the end(norminette)

+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: maeverqu <maeverqu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/23 12:23:23 by maeverqu          #+#    #+#             */
-/*   Updated: 2026/04/23 13:02:56 by maeverqu         ###   ########.fr       */
+/*   Updated: 2026/05/04 16:46:04 by maeverqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	i;
+	int	i;
 
 	i = ft_strlen(s);
 	while (i >= 0)
